requirement.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2021.5.30
charset-normalizer==2.0.6
cmake==3.20.2
colorama==0.4.4
distlib==0.3.2
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
face-recognition-models==0.3.0
filelock==3.0.12
idna==3.2
numpy==1.20.3
Pillow==8.2.0
PyJWT==2.1.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.7
virtualenv==20.4.7
